requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
xlsxwriter>=3.2.0
pandas>=2.2.0
numpy>=1.26.0
streamlit>=1.40.0
python-dateutil>=2.9.0
rapidfuzz>=3.9.0
